move JwsSignerProvider to its own SPI
