+++ settings.gradle.kts
@@ -234,6 +234,7 @@ include(":spi:common:keys-spi")
 include(":spi:common:identity-did-spi")
 include(":spi:common:json-ld-spi")
 include(":spi:common:jwt-spi")
+include(":spi:common:jwt-signer-spi")
 include(":spi:common:token-spi")
 include(":spi:common:oauth2-spi")
 include(":spi:common:policy-engine-spi")
