+++ spi/common/jwt-signer-spi/src/main/java/org/eclipse/edc/jwt/signer/spi/JwsSignerProvider.java
@@ -1,0 +1,25 @@
+/*
+ *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+package org.eclipse.edc.jwt.signer.spi;
+
+import com.nimbusds.jose.JWSSigner;
+
+import java.util.function.Function;
+
+/**
+ * A JwsSignerProvider provides a {@link JWSSigner} for a given private key ID.
+ */
+public interface JwsSignerProvider extends Function<String, JWSSigner> {
+}
